refactor: Move notifiers from pkg to central
