+++ central/notifiers/awssh/mapping_test.go
