+++ central/notifiers/syslog/mocks/syslog.go
