+++ central/notifiers/syslog/syslog_test.go
@@ -8,12 +8,12 @@ import (
 	"testing"
 
 	"github.com/gogo/protobuf/types"
+	"github.com/stackrox/rox/central/notifiers/syslog/mocks"
 	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/features"
 	"github.com/stackrox/rox/pkg/fixtures"
 	metadataGetterMocks "github.com/stackrox/rox/pkg/notifiers/mocks"
-	"github.com/stackrox/rox/pkg/notifiers/syslog/mocks"
 	"github.com/stretchr/testify/suite"
 	"go.uber.org/mock/gomock"
 )
