+++ central/notifiers/syslog/tcp_sender.go
